* Should not using folder where the csproj resides as the project name. If you have a big solution where, for example, unit tests for all project are under unit tests folder, snyk will only check the last project it had scanned.
Since, we don't want to broke any setups, use "--assets-project-name" switch like:

snyk monitor --file=./JobHandlers/VC12.0/JobHandlers.sln --org=coveo-connectors --assets-project-name
