+++ README.md
@@ -8,4 +8,4 @@ Snyk helps you find, fix and monitor for known vulnerabilities in your dependenc
 
 ## Snyk NuGet CLI Plugin
 
-The plugin provides dependency metadata for NuGet projects that manifest dependencies in `project.json` or `packages.config` files.
+The plugin provides dependency metadata for NuGet projects that manifest dependencies in `project.json`, `packages.config` or `project.assets.json` files.
