+++ lib/index.js
@@ -60,6 +60,7 @@ module.exports = {
       root,
       targetFile,
       options.packagesFolder,
-      manifestType).then(createPackageTree);
+      manifestType,
+      options['assets-project-name']).then(createPackageTree);
   },
 };
