+++ lib/nuget-parser/dotnet-core-parser.js
@@ -106,44 +106,9 @@ function getTargetObjToRun(manifest) {
   return manifest.targets[selectedTargetKey];
 }
 
-function validateManifest(manifest) {
-  if (!manifest.project) {
-    throw new Error('Project field was not found in project.assets.json');
-  }
-
-  if (!manifest.project.frameworks) {
-    throw new Error('No frameworks were found in project.assets.json');
-  }
-
-  if (_.isEmpty(manifest.project.frameworks)) {
-    throw new Error('0 frameworks were found in project.assets.json');
-  }
-
-  if (!manifest.targets) {
-    throw new Error('No targets were found in project.assets.json');
-  }
-
-  if (_.isEmpty(manifest.targets)) {
-    throw new Error('0 targets were found in project.assets.json');
-  }
-}
-
 module.exports = {
   parse: function (tree, manifest) {
-    return new Promise(function parseFileContents(resolve, reject) {
-      debug('Trying to parse dot-net-cli manifest');
-
-      try {
-        validateManifest(manifest);
-      } catch (err) {
-        debug('Invalid project.assets.json manifest file');
-        reject(err);
-      }
-
-      if (manifest.project.version) {
-        tree.version = manifest.project.version;
-      }
-
+    return new Promise(function parseFileContents(resolve) {
       // If a targetFramework was not found in the proj file, we will extract it from the lock file
       if (!tree.meta.targetFramework) {
         tree.meta.targetFramework = getFrameworkToRun(manifest);
