+++ lib/nuget-parser/index.js
@@ -8,11 +8,12 @@ const dotnetCoreParser = require('./dotnet-core-parser');
 const dotnetFrameworkParser = require('./dotnet-framework-parser');
 const projectJsonParser = require('./project-json-parser');
 const packagesConfigParser = require('./packages-config-parser');
+const packageReferenceParser = require('./package-reference-parser')
 
 const PARSERS = {
   'dotnet-core': {
     depParser: dotnetCoreParser,
-    fileContentParser: JSON,
+    fileContentParser: packageReferenceParser,
   },
   'project.json': {
     depParser: dotnetFrameworkParser,
@@ -32,7 +33,7 @@ function getPackagesFolder(packagesFolder, projectRootFolder) {
 }
 
 module.exports = {
-  buildDepTreeFromFiles: function (root, targetFile, packagesFolderPath, manifestType) {
+  buildDepTreeFromFiles: function (root, targetFile, packagesFolderPath, manifestType, useProjectNameFromAssetsFile) {
     const fileContentPath = path.resolve(root || '.', targetFile || '.');
     let fileContent;
     try {
@@ -70,7 +71,7 @@ module.exports = {
     const parser = PARSERS[manifestType];
     return parser.depParser.parse(
       tree,
-      parser.fileContentParser.parse(fileContent, tree),
+      parser.fileContentParser.parse(fileContent, tree, useProjectNameFromAssetsFile),
       targetFramework,
       packagesFolder);
   },
